Discretized the single humongous AD.h into separate .h files.
