+++ AD-Tool/AD-Tool/AD-Tool.vcxproj
@@ -110,6 +110,7 @@
     <Link>
       <SubSystem>Console</SubSystem>
       <GenerateDebugInformation>true</GenerateDebugInformation>
+      <AdditionalDependencies>%(AdditionalDependencies)</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
@@ -155,7 +156,14 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="AD.h" />
+    <ClInclude Include="counter.h" />
+    <ClInclude Include="gNode.h" />
+    <ClInclude Include="OP.h" />
+    <ClInclude Include="scalar.h" />
     <ClInclude Include="matrix.h" />
+    <ClInclude Include="node.h" />
+    <ClInclude Include="VAL.h" />
+    <ClInclude Include="var.h" />
     <ClInclude Include="vector.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
