+++ AD-Tool/AD-Tool/AD.h
@@ -1,399 +1,24 @@
 /** Make changes in the code to take care of intermediate variable inclusion!! */
-#ifndef AD_H
-#define AD_H
+#pragma once
 
 #include <iostream>
 #include <cmath>
+
 #include "vector.h"
 #include "matrix.h"
 
-static int idc = 0;
+#include "node.h"
+#include "scalar.h"
+#include "var.h"
+#include "OP.h"
+#include "gNode.h"
+
+#include "VAL.h"
+
+#include "counter.h"
 
 using namespace std;
 
-template <class T>
-class VAL;
-
-template<class T>
-class AD;
-
-/**CLASS: node ***************************************/
-template<class T>
-class node {
-	node* Lchild, * Rchild;
-
-public:
-	node() { Lchild = NULL; Rchild = NULL; }
-	virtual VAL<T> getVAL() = 0;
-	virtual node<T>* getself() = 0;
-	virtual void disp() = 0;
-	virtual char gettype() = 0;
-	virtual node<T>* clone() = 0;
-
-	void InOrderTreeWalk();
-
-	void setLchild(node* p) { this->Lchild = p; }
-	void setRchild(node* p) { this->Rchild = p; }
-
-	node<T>* getLchild() { return this->Lchild; }
-	node<T>* getRchild() { return this->Rchild; }
-
-	VAL<T> getleftVAL() { return Lchild->getVAL(); }
-	VAL<T> getrightVAL() { return Rchild->getVAL(); }
-};
-
-template<class T>
-void node<T>::InOrderTreeWalk() {
-	if (this->Lchild != NULL) {
-		cout << "LEFT: ";
-		this->Lchild->InOrderTreeWalk();
-	}
-
-	cout << "THIS ";
-	cout << " [" << this->gettype() << "] : ";
-	this->disp();
-
-	if (this->Rchild != NULL) {
-		cout << "RGHT: ";
-		this->Rchild->InOrderTreeWalk();
-	}
-}
-
-/**CLASS: scalar ***************************************/
-template<class T>
-class scalar :public node<T> {
-	T val;
-	scalar<T>* self;
-	char type;
-
-public:
-	scalar() { this->self = this; this->type = 's'; }
-	scalar(T k) { this->val = k; this->self = this; this->type = 's'; }
-
-	T getval() { return val; }
-
-	void operator =(T k) { this->val = k; }
-	void operator =(scalar<T> s) { this->val = s.val; }
-
-	void disp() {
-		cout << "SCALAR: val = " << this->val << ", address = " << this->self << endl;
-	}
-	node<T>* clone() { return this; }
-	scalar<T>* getself() { return this->self; }
-	char gettype() { return this->type; }
-	VAL<T> getVAL();
-};
-
-template<class T>
-VAL<T> scalar<T>::getVAL() {
-	VAL<T> v;
-	v.input_f(this->val);
-	T* df = new T[idc];
-
-	for (int i = 0; i < idc; i++)
-		df[i] = 0.0;
-
-	v.input_df(df);
-
-	return v;
-}
-
-
-/**CLASS: var ***************************************/
-template<class T>
-class var :public node<T> {
-	int id;
-	T val;
-	var<T>* self;
-	char type;
-
-public:
-	var() { this->id = idc++; this->self = this; this->type = 'v'; }
-	var(T k) { this->id = idc++; this->val = k; this->self = this; this->type = 'v'; }
-
-	void operator =(T k) { this->val = k; }
-	void operator =(scalar<T> s) { this->p = s.getval(); }
-	void operator =(var<T> x) { this->val = x.val; }
-
-	void disp() {
-		cout << "VAR: id = " << this->id << ", val = " << this->val << ", address = " << this->self << endl;
-	}
-	node<T>* clone() { return this; }
-	var<T>* getself() { return this->self; }
-	char gettype() { return this->type; }
-	VAL<T> getVAL();
-};
-
-template<class T>
-VAL<T> var<T>::getVAL() {
-	VAL<T> v;
-	v.input_f(this->val);
-	T* df = new T[idc];
-
-	for (int i = 0; i < this->id; i++)
-		df[i] = 0.0;
-
-	df[this->id] = 1.0;
-
-	for (int i = this->id + 1; i < idc; i++)
-		df[i] = 0.0;
-
-	v.input_df(df);
-
-	return v;
-}
-
-
-/**CLASS: OP ***************************************/
-template<class T>
-class OP :public node<T> {
-	OP* self;
-	char op, type;
-
-public:
-	OP() { this->op = '\0'; this->self = this; this->type = 'o'; }
-	OP(char ch) { this->op = ch; this->self = this; this->type = 'o'; }
-	OP(char ch, node<T>* L) { this->op = ch; this->setLchild(L); this->self = this; this->type = 'o'; }
-	OP(char ch, node<T>* L, node<T>* R) { this->op = ch; this->setLchild(L); this->setRchild(R); this->self = this; this->type = 'o'; }
-	OP(OP<T>& O) { this->op = O.op; this->setLchild(O.getLchild()); this->setRchild(O.getRchild()); this->self = this; this->type = 'o'; }
-
-	void operator =(OP<T> O) { this->op = O.op; this->setLchild(O.getLchild()); this->setRchild(O.getRchild()); }
-
-	OP<T>* getself() { return this->self; }
-	char gettype() { return this->type; }
-
-	node<T>* clone() { return new OP<T>(*this); }
-
-	void disp() {
-		cout << "OP: op = " << this->op << ", address = " << this->self << endl;
-	}
-
-	VAL<T> getVAL();
-};
-
-template<class T>
-VAL<T> OP<T>::getVAL() {
-	VAL<T> v;
-	switch (this->op) {
-	case '+':
-		if (this->getRchild() == NULL)
-			v = this->getleftVAL();
-		else
-			v = this->getleftVAL() + this->getrightVAL();
-		break;
-	case '-':
-		if (this->getRchild() == NULL)
-			v = -(this->getleftVAL());
-		else
-			v = this->getleftVAL() - this->getrightVAL();
-		break;
-	case '*': v = this->getleftVAL() * this->getrightVAL(); break;
-	case '/': v = this->getleftVAL() / this->getrightVAL(); break;
-	case '^': v = this->getleftVAL() ^ this->getrightVAL(); break;
-	default: cout << "ERROR: Invalid Operator.\n"; exit(1); break;
-	}
-	return v;
-}
-
-
-/**CLASS: gNode ***************************************/
-template<class T>
-class gNode :public node<T> {
-	node<T>* p, * self;
-	char type;
-public:
-	gNode() { this->p = NULL; this->self = this; this->type = 'g'; }
-	gNode(node<T>* P) { this->p = P; this->self = this; this->type = 'g'; }
-
-	void operator =(T k) { this->p = new scalar<T>(k); }
-	void operator =(scalar<T> s) { this->p = s.getself(); }
-	void operator =(var<T> x) { this->p = x.getself(); }
-	void operator =(OP<T> O) { this->p = O.getself(); }
-	void operator =(AD<T> G) { this->p = G.geteval(); }
-
-	void setNodePointer(node<T>* new_p) { this->p = new_p; }
-	node<T>* getNodePointer() { return this->p; }
-
-	node<T>* clone() { return this->p; }
-	VAL<T> getVAL() { return this->p->getVAL(); }
-	node<T>* getself() { return this->self; }
-	void disp() { this->p->disp(); }
-	char gettype() { return this->type; }
-};
-
-
-/**CLASS: VAL ***************************************/
-template <class T>
-class VAL
-{
-	T f;
-	T* df;
-public:
-	VAL();
-
-	void disp() {
-		cout << "f = " << this->f << endl;
-		cout << "df = ";
-		for (int i = 0; i < idc; i++)
-			cout << this->df[i] << "   ";
-		cout << endl;
-	}
-
-	void input_f(T k) { this->f = k; }
-	void input_df(T* dk) { this->df = dk; }
-
-	T getf() { return f; }
-
-	T& operator [] (int);
-
-	/**Member Functions*/
-	/**Unary operators*/
-	VAL operator +(); //defined
-	VAL operator -(); //defined
-
-	/**Binary operators*/
-	void operator =(VAL); // defined
-	VAL operator +(VAL); // defined
-	VAL operator -(VAL); // defined
-	VAL operator *(VAL); // defined
-	VAL operator /(VAL); // defined
-	VAL operator ^(VAL); // defined
-};
-
-template<class T>
-VAL<T>::VAL() {
-	this->f = 0.0;
-	this->df = new T[idc];
-
-	for (int i = 0; i < idc; i++)
-		this->df[i] = 0.0;
-}
-
-template<class T>
-T& VAL<T>::operator [] (int k) {
-	if (k < idc)
-		return df[k];
-	else {
-		cout << "Index out of bound!";
-		exit(1);
-	}
-}
-
-/**VAL: Operator Overloading*/
-template<class T>
-void VAL<T>::operator =(VAL g) {
-	this->f = g.f;
-	for (int i = 0; i < idc; i++)
-		this->df[i] = g.df[i];
-}
-
-template<class T>
-VAL<T> VAL<T>::operator +() {
-	VAL<T> h;
-	h.f = f;
-	for (int i = 0; i < idc; i++)
-		h.df[i] = this->df[i];
-	return h;
-}
-
-template<class T>
-VAL<T> VAL<T>::operator -() {
-	VAL<T> h;
-	h.f = -(this->f);
-	for (int i = 0; i < idc; i++)
-		h.df[i] = -(this->df[i]);
-	return h;
-}
-
-template<class T>
-VAL<T> VAL<T>::operator +(VAL g) {
-	VAL<T> h;
-	h.f = this->f + g.f;
-	for (int i = 0; i < idc; i++)
-		h.df[i] = this->df[i] + g.df[i];
-	return h;
-}
-
-template<class T>
-VAL<T> VAL<T>::operator -(VAL g) {
-	VAL<T> h;
-	h.f = this->f - g.f;
-	for (int i = 0; i < idc; i++)
-		h.df[i] = this->df[i] - g.df[i];
-	return h;
-}
-
-template<class T>
-VAL<T> VAL<T>::operator *(VAL g) {
-	VAL<T> h;
-	h.f = this->f * g.f;
-	for (int i = 0; i < idc; i++)
-		h.df[i] = g.f * (this->df[i]) + this->f * g.df[i];
-	return h;
-}
-
-template<class T>
-VAL<T> VAL<T>::operator /(VAL g) {
-	if (g.f == 0) {
-		cout << "MATH ERROR: Division by Zero!\n";
-		exit(1);
-	}
-	else {
-		VAL<T> h;
-		h.f = this->f / g.f;
-		for (int i = 0; i < idc; i++)
-			h.df[i] = this->df[i] / g.f - this->f * g.df[i] / (g.f * g.f);
-		return h;
-	}
-}
-
-template<class T>
-VAL<T> VAL<T>::operator ^(VAL g) {
-	VAL<T> h;
-	if (this->f == 0 && g.f == 0) {
-		cout << "MATH ERROR: 0^0 Undefined!\n";
-		exit(1);
-	}
-	else if (this->f == 0 && g.f < 0) {
-		cout << "MATH ERROR: Division by Zero!\n";
-		exit(1);
-	}
-	else
-		h.f = pow(this->f, g.f);
-
-	int g_df = 0;
-
-	for (int i = 0; i < idc && g_df == 0; i++)
-		if (g.df[i] != 0)
-			g_df = 1;
-
-	if (g_df == 0) {
-		if (this->f != 0) {
-			for (int i = 0; i < idc; i++)
-				h.df[i] = pow(this->f, (g.f - 1)) * g.f * this->df[i];
-		}
-		else if (g.f > 1) {
-			for (int i = 0; i < idc; i++)
-				h.df[i] = 0.0;
-		}
-		else {
-			cout << "MATH ERROR: Division by Zero!\n";
-			exit(1);
-		}
-	}
-	else if (this->f > 0) {
-		for (int i = 0; i < idc; i++)
-			h.df[i] = pow(this->f, (g.f - 1)) * g.f * this->df[i] + pow(this->f, g.f) * g.df[i] * log(this->f);
-	}
-	else {
-		cout << "MATH ERROR: LOG undefined for values <= 0!\n";
-		exit(1);
-	}
-	return h;
-}
-
-
-/**CLASS: AD ***************************************/
 template<class T>
 class AD {
 	VAL<T> val;
@@ -599,16 +224,23 @@ AD<T> copyAD(AD<T> G) {
 
 template<class T>
 node<T>* cloneNode(node<T>* root) {
+
 	switch (root->gettype()) {
-	case 's': return root; break;
-	case 'v': return root; break;
-	case 'g': return cloneNode(root->clone()); break;
-	case 'o':
+	case 's': { return root; break; }
+	case 'v': { return root; break; }
+	case 'g': { return cloneNode(root->clone()); break; }
+	case 'o': {
 		node<T>* O = root->clone();
 		O->setLchild(cloneNode(O->getLchild()));
 		O->setRchild(cloneNode(O->getRchild()));
 		return O;
 		break;
+	}
+	default: {
+		cout << "Invalid node object.\n";
+		return NULL;
+		break;
+	}
 	}
 }
 
@@ -1676,19 +1308,21 @@ AD<T> operator ^(gNode<T> g, gNode<T> h) {
 /**Advanced operations*/
 template<class T>
 vector<T> grad(AD<T> G) {
-	vector<T> V(idc);
+	counter idc;
+	vector<T> V(idc.getCount());
 	for (int i = 0; i < idc; i++)
 		V[i] = G.val[i];
 
 	return V;
 }
+
 template<class T>
 matrix<T> jacobian(int n, AD<T>* AD_list) {
-	matrix<T> M(n, idc);
+	counter idc;
+	matrix<T> M(n, idc.getCount());
 	for (int i = 0; i < n; i++)
-		for (int j = 0; j < idc; j++)
+		for (int j = 0; j < idc ; j++)
 			M(i, j) = AD_list[i].val[j];
 
 	return M;
 }
-#endif
