--- AD-Tool/AD-Tool/Debug/AD-Tool.log
@@ -1,4 +1,0 @@
-﻿  AD-test.cpp
-D:\nsr\my stuff\self projects\AD-Tool (Visual Studio)\AD-Tool\AD-Tool\AD-test.cpp(7,16): warning C4101:  'p': unreferenced local variable
-D:\nsr\my stuff\self projects\AD-Tool (Visual Studio)\AD-Tool\AD-Tool\AD-test.cpp(7,16): warning C4101: D:\nsr\my stuff\self projects\AD-Tool (Visual Studio)\AD-Tool\AD-Tool\AD.h(613) : warning C4715: 'cloneNode<double>': not all control paths return a value
-  AD-Tool.vcxproj -> D:\nsr\my stuff\self projects\AD-Tool (Visual Studio)\AD-Tool\Debug\AD-Tool.exe
