--- AD-Tool/AD-Tool/Debug/AD-Tool.tlog/AD-Tool.lastbuildstate
@@ -1,2 +1,0 @@
-#TargetFrameworkVersion=v4.0:PlatformToolSet=v142:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0
-Debug|Win32|D:\nsr\my stuff\self projects\AD-Tool (Visual Studio)\AD-Tool\|
